pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gallerator"
description = "Generate image galleries"
readme = "README.md"
version = "1.0"
dependencies = [
  "Pillow",
  "vcsi",
  "tqdm",
  "Jinja2",
  "ruff",
]
# For Path.relative_to(other, walk_up=True)
requires-python = ">= 3.12"

[project.urls]
Homepage = "https://github.com/pmorch/gallerator"
Demos = "https://pmorch.github.io/demo-gallerator/"

[project.scripts]
gallerator = "gallerator.main:cli_main"
